// frontend/taskhub-web/src/app/api/generated/model/workItemDto.ts
/**
 * TaskHub.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkItemStatus } from './workItemStatus';
import { Priority } from './priority';


export interface WorkItemDto { 
    id?: number;
    title?: string | null;
    description?: string | null;
    status?: WorkItemStatus;
    priority?: Priority;
    dueDate?: string | null;
    createdAt?: string;
}