// frontend/taskhub-web/src/app/api/generated/model/workItemStatus.ts
/**
 * TaskHub.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum WorkItemStatus {

    New = 'New',

    InProgress = 'InProgress',

    Done = 'Done'
}